test(settings): follow the Email Accounts config/list/form split in the suites
Hand-applied from 0cb8ba47d's two Email-Accounts hunks (not cherry-picked
whole -- that commit also touches ~30 unrelated files). Points
emailAccountDefaults.test.ts/emailConfigValidate.test.ts at emailAccounts.ts
instead of the deleted emailConfig.ts, and stubs frappe-ui/@/utils/dialogs/
useSettingsListResource, which the config module now also pulls in for the
list's row actions.
